[SC-163262] Suppress diff when value_type doesn't change
Because of how Terraform defaults work, Terraform will attempt to store
a `value_type` for `values` even when `values` is empty. This adds a
[DiffSuppressFunc](https://www.terraform.io/plugin/sdkv2/schemas/schema-behaviors#diffsuppressfunc)
for the `value_type` in the schema that checks if the diff went from
empty `""` (what's in the TF manifest) to `"string"` (the default
value) and if there is no `value_type` in the resource (so the
`value_type` wasn't actually updated by the user) then we can safely
ignore the diff. This suppresses annoying messages for users without
impacting the behavior of the provider.