Retry failed acceptance tests once
This modifies the GNUMakefile `testacc` target to run the specified
acceptance tests once, and if they fail retry them one time. This is an
effort to prevent alerts for flaky tests in the channel, which we've
been seeing an increase in. Because we run each test individually in CI,
this is actually pretty nice since just the tests that fail will get
rerun and not all tests (unless running all tests locally). We may at
some point want to add an environment variable or other configuration
around this.